Did some work to separate out some of the user interface code for LookbackMultiple and the calculation portion of it.  These now exist in different modules.  Did some work to try to optimize and improve the speed of the LookbackMultiple functionality.  Found an issue with pytz.timezone() taking a long time on the first call (see README.txt notes on this).  Found out through profiling and testing that the Ephemeris execution time does not change very much by getting rid of equatorial coordinate calculations and getting rid of rectangular coordinate calculations.
