+++ conf/logging.conf
@@ -3,7 +3,7 @@
 ##############################################################################
 
 [loggers]
-keys=root,astrologychart,data_objects,dialogs,ephemeris,geonames,lookbackmultiple,main,pricebarchart,pricebarchart_dialogs,pricebarspreadsheet,spreadsheet_calc,ui,widgets
+keys=root,astrologychart,data_objects,dialogs,ephemeris,geonames,lookbackmultiple_calc,lookbackmultiple_ui,main,pricebarchart,pricebarchart_dialogs,pricebarspreadsheet,spreadsheet_calc,ui,widgets
 
 [handlers]
 keys=consoleHandler,rotatingFileHandler,fileHandler
@@ -57,12 +57,19 @@ handlers=rotatingFileHandler
 propagate=1
 qualname=geonames
 
-[logger_lookbackmultiple]
+[logger_lookbackmultiple_calc]
 #level=DEBUG
 level=INFO
 handlers=rotatingFileHandler
 propagate=1
-qualname=lookbackmultiple
+qualname=lookbackmultiple_calc
+
+[logger_lookbackmultiple_ui]
+#level=DEBUG
+level=INFO
+handlers=rotatingFileHandler
+propagate=1
+qualname=lookbackmultiple_ui
 
 [logger_main]
 #level=DEBUG
