+++ src/lookbackmultiple_calc.py
@@ -1,38 +1,30 @@
 
 
-# For logging.
-import logging
-
-# For newlines.
+# For directory access.
 import os
-
-# For printing out the current method's name.
+import sys
 import inspect
-
-# For copy.deepcopy.
-import copy
 
 # For timestamps and timezone information.
 import datetime
 import pytz
 
-# For PyQt UI classes.
-from PyQt4 import QtCore
-from PyQt4.QtCore import *
-from PyQt4.QtGui import *
+# For math.floor()
+import math
 
-# Import image resources.
-import resources
+# For logging.
+import logging
+import logging.config
 
-from widgets import ColorIcon
+# Import the Swiss Ephemeris
+import swisseph as swe
+
+from ephemeris import PlanetaryInfo
+from ephemeris import Ephemeris
 
 from data_objects import Util
-from data_objects import LookbackMultiple
 
-from dialogs import LookbackMultipleEditDialog
-
-# For calculating LookbackMultiple periods using planets.
-from ephemeris import Ephemeris
+##############################################################################
 
 class LookbackMultipleUtils:
     """Contains various static methods to assist in calculating 
@@ -40,7 +32,7 @@ class LookbackMultipleUtils:
     """
     
     # Logger object for this class.
-    log = logging.getLogger("lookbackmultiple.LookbackMultipleUtils")
+    log = logging.getLogger("lookbackmultiple_calc.LookbackMultipleUtils")
     
     @staticmethod
     def getDatetimesOfLongitudeDeltaDegreesInFuture(\
@@ -93,7 +85,8 @@ class LookbackMultipleUtils:
         'referenceDt'.
         """
         
-        LookbackMultipleUtils.log.debug("Entered " + inspect.stack()[0][3] + "()")
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Entered " + inspect.stack()[0][3] + "()")
 
         # Return value.
         rv = []
@@ -146,8 +139,8 @@ class LookbackMultipleUtils:
 
         # Iterate through, creating artfacts and adding them as we go.
         steps = []
-        steps.append(copy.deepcopy(referenceDt))
-        steps.append(copy.deepcopy(referenceDt))
+        steps.append(referenceDt)
+        steps.append(referenceDt)
 
         longitudesP1 = []
         longitudesP1.append(None)
@@ -169,7 +162,8 @@ class LookbackMultipleUtils:
 
             return fieldValue
             
-        LookbackMultipleUtils.log.debug("Stepping through timestamps from {} ...".\
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Stepping through timestamps from {} ...".\
                   format(Ephemeris.datetimeToStr(referenceDt)))
 
         currDiff = None
@@ -184,19 +178,23 @@ class LookbackMultipleUtils:
             currDt = steps[-1]
             prevDt = steps[-2]
             
-            LookbackMultipleUtils.log.debug("Looking at currDt == {} ...".\
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("Looking at currDt == {} ...".\
                       format(Ephemeris.datetimeToStr(currDt)))
 
             p1 = Ephemeris.getPlanetaryInfo(planetName, currDt)
 
             if planetReferenceLongitude == None:
                 planetReferenceLongitude = getFieldValue(p1, fieldName)
-                LookbackMultipleUtils.log.debug("planetReferenceLongitude == {}".\
+
+                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                    LookbackMultipleUtils.log.debug("planetReferenceLongitude == {}".\
                                                 format(planetReferenceLongitude))
             
             longitudesP1[-1] = getFieldValue(p1, fieldName)
             
-            LookbackMultipleUtils.log.debug("{} {} {} {} is: {}".\
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("{} {} {} {} is: {}".\
                       format(p1.name, centricityType, longitudeType, fieldName,
                              getFieldValue(p1, fieldName)))
             
@@ -207,15 +205,17 @@ class LookbackMultipleUtils:
             currDiff = Util.toNormalizedAngle(\
                 longitudesP1[-1] - planetReferenceLongitude)
             
-            LookbackMultipleUtils.log.debug("prevDiff == {}".format(prevDiff))
-            LookbackMultipleUtils.log.debug("currDiff == {}".format(currDiff))
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("prevDiff == {}".format(prevDiff))
+                LookbackMultipleUtils.log.debug("currDiff == {}".format(currDiff))
             
             # If this is not the first iteration of the loop 
             # (i.e. we have both a prevDiff and a currDiff to compare).
             if prevDiff != None and longitudesP1[-2] != None:
                 
                 if prevDiff > 240 and currDiff < 120:
-                    LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
                               format(planetReferenceLongitude) + \
                               "from below to above!")
 
@@ -226,7 +226,8 @@ class LookbackMultipleUtils:
                     
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -264,7 +265,8 @@ class LookbackMultipleUtils:
                     numFullCircles += 1
 
                 elif prevDiff < 120 and currDiff > 240:
-                    LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
                               format(planetReferenceLongitude) + \
                               "from above to below!")
 
@@ -275,7 +277,8 @@ class LookbackMultipleUtils:
 
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -315,8 +318,9 @@ class LookbackMultipleUtils:
                 # Calculate the total number of degrees elapsed so far.
                 currElapsed = (numFullCircles * 360.0) + currDiff
 
-                LookbackMultipleUtils.log.debug("currElapsed == {}".format(currElapsed))
-                LookbackMultipleUtils.log.debug("desiredDeltaDegrees == {}".\
+                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                    LookbackMultipleUtils.log.debug("currElapsed == {}".format(currElapsed))
+                    LookbackMultipleUtils.log.debug("desiredDeltaDegrees == {}".\
                           format(desiredDeltaDegrees))
                 
                 if currElapsed > desiredDeltaDegrees:
@@ -326,7 +330,8 @@ class LookbackMultipleUtils:
                     # moments in time where the planet is elapsed this
                     # many degrees (in the event that the planet goes
                     # retrograde).
-                    LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                               "elapsed degrees from below to above.  " + \
                               "Narrowing down to the exact moment in time ...")
                     
@@ -336,7 +341,8 @@ class LookbackMultipleUtils:
                         planetReferenceLongitude + 
                         (desiredDeltaDegrees % 360.0))
 
-                    LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
                     
                     # Check starting from steps[-2] to steps[-1] to
                     # see exactly when it passes this desiredDegree.
@@ -348,7 +354,8 @@ class LookbackMultipleUtils:
                     
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -378,7 +385,8 @@ class LookbackMultipleUtils:
                     # t2 holds the moment in time.
                     rv.append(t2)
 
-                    LookbackMultipleUtils.log.debug("First moment in time found to be: {}".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("First moment in time found to be: {}".\
                               format(Ephemeris.datetimeToStr(t2)))
 
                     # Now we will want to find the other elapsed points, if they
@@ -392,7 +400,8 @@ class LookbackMultipleUtils:
                         getFieldValue(p1, fieldName) - desiredDegree)
                     currDiff = None
 
-                    LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
                     
                     # There is only need to continue looking for more potential
                     # timestamps if the planet can go retrograde.  Direct-only
@@ -406,13 +415,14 @@ class LookbackMultipleUtils:
                           Ephemeris.isAscmcPlanetName(planetName) or
                           planetName == "MoSu")):
                          
-                         LookbackMultipleUtils.log.debug(\
-                             "No need to look for anymore timestamps " + \
-                             "because this planet doesn't go retrograde.")
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug(\
+                            "No need to look for anymore timestamps " + \
+                            "because this planet doesn't go retrograde.")
 
-                         # Set the done flag, which will stop us from looking
-                         # for more timestamps.
-                         done = True
+                        # Set the done flag, which will stop us from looking
+                        # for more timestamps.
+                        done = True
                     
                     while (prevDiff <= 120 or prevDiff > 240) and done != True:
 
@@ -420,7 +430,8 @@ class LookbackMultipleUtils:
                         currDiff = Util.toNormalizedAngle(\
                             getFieldValue(p1, fieldName) - desiredDegree)
 
-                        LookbackMultipleUtils.log.debug("currDt == {}, ".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("currDt == {}, ".\
                                   format(Ephemeris.datetimeToStr(currDt)) + 
                                   "longitude == {}, ".\
                                   format(getFieldValue(p1, fieldName)) + \
@@ -428,7 +439,8 @@ class LookbackMultipleUtils:
                                   format(currDiff))
 
                         if prevDiff > 240 and currDiff < 120:
-                            LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                                       "elapsed degrees from " + \
                                       "below to above.  " + \
                                       "Narrowing down to the exact moment " + \
@@ -442,7 +454,8 @@ class LookbackMultipleUtils:
                             # Refine the timestamp until it is less
                             # than the threshold.
                             while currErrorTd > maxErrorTd:
-                                LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                    LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                           format(Ephemeris.datetimeToStr(t1),
                                                  Ephemeris.datetimeToStr(t2)))
                                 
@@ -476,12 +489,14 @@ class LookbackMultipleUtils:
 
 
                             # currDt holds the moment in time.
-                            LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
                                       format(Ephemeris.datetimeToStr(currDt)))
                             rv.append(currDt)
 
                         elif prevDiff < 120 and currDiff > 240:
-                            LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                                       "elapsed degrees from " + \
                                       "above to below.  " + \
                                       "Narrowing down to the exact moment " + \
@@ -495,7 +510,8 @@ class LookbackMultipleUtils:
                             # Refine the timestamp until it is less
                             # than the threshold.
                             while currErrorTd > maxErrorTd:
-                                LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                    LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                           format(Ephemeris.datetimeToStr(t1),
                                                  Ephemeris.datetimeToStr(t2)))
                                 
@@ -528,22 +544,25 @@ class LookbackMultipleUtils:
                                 currErrorTd = Util.absTd(t2 - t1)
         
                             # currDt holds the moment in time.
-                            LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
                                       format(Ephemeris.datetimeToStr(currDt)))
+
                             rv.append(currDt)
 
                         prevDt = currDt
-                        currDt = copy.deepcopy(currDt) + stepSizeTd
+                        currDt = currDt + stepSizeTd
                         prevDiff = currDiff
                         currDiff = None
 
-                    LookbackMultipleUtils.log.debug("Done searching for timestamps.")
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Done searching for timestamps.")
                         
                     # We have our timestamps, so we are done.
                     done = True
                     
             # Prepare for the next iteration.
-            steps.append(copy.deepcopy(steps[-1]) + stepSizeTd)
+            steps.append(steps[-1] + stepSizeTd)
             del steps[0]
             longitudesP1.append(None)
             del longitudesP1[0]
@@ -551,7 +570,9 @@ class LookbackMultipleUtils:
             # Update prevDiff as the currDiff.
             prevDiff = currDiff
 
-        LookbackMultipleUtils.log.debug("Exiting " + inspect.stack()[0][3] + "()")
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Exiting " + inspect.stack()[0][3] + "()")
+
         return rv
 
         
@@ -606,7 +627,8 @@ class LookbackMultipleUtils:
         'referenceDt'.
         """
         
-        LookbackMultipleUtils.log.debug("Entered " + inspect.stack()[0][3] + "()")
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Entered " + inspect.stack()[0][3] + "()")
 
         # Return value.
         rv = []
@@ -662,8 +684,8 @@ class LookbackMultipleUtils:
 
         # Iterate through, creating artfacts and adding them as we go.
         steps = []
-        steps.append(copy.deepcopy(referenceDt))
-        steps.append(copy.deepcopy(referenceDt))
+        steps.append(referenceDt)
+        steps.append(referenceDt)
 
         longitudesP1 = []
         longitudesP1.append(None)
@@ -685,7 +707,8 @@ class LookbackMultipleUtils:
 
             return fieldValue
             
-        LookbackMultipleUtils.log.debug("Stepping through timestamps from {} ...".\
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Stepping through timestamps from {} ...".\
                   format(Ephemeris.datetimeToStr(referenceDt)))
 
         currDiff = None
@@ -700,19 +723,23 @@ class LookbackMultipleUtils:
             currDt = steps[-1]
             prevDt = steps[-2]
             
-            LookbackMultipleUtils.log.debug("Looking at currDt == {} ...".\
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("Looking at currDt == {} ...".\
                       format(Ephemeris.datetimeToStr(currDt)))
 
             p1 = Ephemeris.getPlanetaryInfo(planetName, currDt)
 
             if planetReferenceLongitude == None:
                 planetReferenceLongitude = getFieldValue(p1, fieldName)
-                LookbackMultipleUtils.log.debug("planetReferenceLongitude == {}".\
+
+                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                    LookbackMultipleUtils.log.debug("planetReferenceLongitude == {}".\
                                                 format(planetReferenceLongitude))
 
             longitudesP1[-1] = getFieldValue(p1, fieldName)
             
-            LookbackMultipleUtils.log.debug("{} {} {} {} is: {}".\
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("{} {} {} {} is: {}".\
                       format(p1.name, centricityType, longitudeType, fieldName,
                              getFieldValue(p1, fieldName)))
             
@@ -723,15 +750,17 @@ class LookbackMultipleUtils:
             currDiff = Util.toNormalizedAngle(\
                 longitudesP1[-1] - planetReferenceLongitude)
 
-            LookbackMultipleUtils.log.debug("prevDiff == {}".format(prevDiff))
-            LookbackMultipleUtils.log.debug("currDiff == {}".format(currDiff))
+            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                LookbackMultipleUtils.log.debug("prevDiff == {}".format(prevDiff))
+                LookbackMultipleUtils.log.debug("currDiff == {}".format(currDiff))
             
             # If this is not the first iteration of the loop 
             # (i.e. we have both a prevDiff and a currDiff to compare).
             if prevDiff != None and longitudesP1[-2] != None:
                 
                 if prevDiff > 240 and currDiff < 120:
-                    LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
                               format(planetReferenceLongitude) + \
                               "from below to above!")
 
@@ -742,7 +771,8 @@ class LookbackMultipleUtils:
                     
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -780,7 +810,8 @@ class LookbackMultipleUtils:
                     numFullCircles += 1
 
                 elif prevDiff < 120 and currDiff > 240:
-                    LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Crossed over planet reference longitude {} ".\
                               format(planetReferenceLongitude) + \
                               "from above to below!")
 
@@ -791,7 +822,8 @@ class LookbackMultipleUtils:
 
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -820,7 +852,8 @@ class LookbackMultipleUtils:
                             currDt = t2
                             currDiff = testDiff
 
-                            LookbackMultipleUtils.log.debug("currDiff updated to: {}".format(currDiff))
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("currDiff updated to: {}".format(currDiff))
 
                         currErrorTd = Util.absTd(t2 - t1)
 
@@ -833,8 +866,9 @@ class LookbackMultipleUtils:
                 # Calculate the total number of degrees elapsed so far.
                 currElapsed = (numFullCircles * 360.0) + currDiff
 
-                LookbackMultipleUtils.log.debug("currElapsed == {}".format(currElapsed))
-                LookbackMultipleUtils.log.debug("desiredDeltaDegrees == {}".\
+                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                    LookbackMultipleUtils.log.debug("currElapsed == {}".format(currElapsed))
+                    LookbackMultipleUtils.log.debug("desiredDeltaDegrees == {}".\
                           format(desiredDeltaDegrees))
                 
                 if currElapsed < desiredDeltaDegrees:
@@ -844,7 +878,8 @@ class LookbackMultipleUtils:
                     # moments in time where the planet is elapsed this
                     # many degrees (in the event that the planet goes
                     # retrograde).
-                    LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                               "elapsed degrees from above to below.  " + \
                               "Narrowing down to the exact moment in time ...")
                     
@@ -854,7 +889,8 @@ class LookbackMultipleUtils:
                         planetReferenceLongitude + 
                         (desiredDeltaDegrees % 360.0))
 
-                    LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
                     
                     # Check starting from steps[-2] to steps[-1] to
                     # see exactly when it passes this desiredDegree.
@@ -866,7 +902,8 @@ class LookbackMultipleUtils:
                     
                     # Refine the timestamp until it is less than the threshold.
                     while currErrorTd > maxErrorTd:
-                        LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                   format(Ephemeris.datetimeToStr(t1),
                                          Ephemeris.datetimeToStr(t2)))
 
@@ -896,7 +933,8 @@ class LookbackMultipleUtils:
                     # t2 holds the moment in time.
                     rv.append(t2)
 
-                    LookbackMultipleUtils.log.debug("First moment in time found to be: {}".\
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("First moment in time found to be: {}".\
                               format(Ephemeris.datetimeToStr(t2)))
 
                     # Now we will want to find the other elapsed points, if they
@@ -910,7 +948,8 @@ class LookbackMultipleUtils:
                         getFieldValue(p1, fieldName) - desiredDegree)
                     currDiff = None
 
-                    LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("desiredDegree == {}".format(desiredDegree))
                     
                     # There is only need to continue looking for more potential
                     # timestamps if the planet can go retrograde.  Direct-only
@@ -924,13 +963,14 @@ class LookbackMultipleUtils:
                           Ephemeris.isAscmcPlanetName(planetName) or
                           planetName == "MoSu")):
                          
-                         LookbackMultipleUtils.log.debug(\
-                             "No need to look for anymore timestamps " + \
-                             "because this planet doesn't go retrograde.")
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug(\
+                            "No need to look for anymore timestamps " + \
+                            "because this planet doesn't go retrograde.")
 
-                         # Set the done flag, which will stop us from looking
-                         # for more timestamps.
-                         done = True
+                        # Set the done flag, which will stop us from looking
+                        # for more timestamps.
+                        done = True
                     
                     while (prevDiff >= 240 or prevDiff < 120) and done != True:
 
@@ -938,7 +978,8 @@ class LookbackMultipleUtils:
                         currDiff = Util.toNormalizedAngle(\
                             getFieldValue(p1, fieldName) - desiredDegree)
 
-                        LookbackMultipleUtils.log.debug("currDt == {}, ".\
+                        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                            LookbackMultipleUtils.log.debug("currDt == {}, ".\
                                   format(Ephemeris.datetimeToStr(currDt)) + 
                                   "longitude == {}, ".\
                                   format(getFieldValue(p1, fieldName)) + \
@@ -946,7 +987,8 @@ class LookbackMultipleUtils:
                                   format(currDiff))
 
                         if prevDiff < 120 and currDiff > 240:
-                            LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                                       "elapsed degrees from " + \
                                       "below to above.  " + \
                                       "Narrowing down to the exact moment " + \
@@ -960,7 +1002,8 @@ class LookbackMultipleUtils:
                             # Refine the timestamp until it is less
                             # than the threshold.
                             while currErrorTd > maxErrorTd:
-                                LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                    LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                           format(Ephemeris.datetimeToStr(t1),
                                                  Ephemeris.datetimeToStr(t2)))
                                 
@@ -994,12 +1037,14 @@ class LookbackMultipleUtils:
 
 
                             # currDt holds the moment in time.
-                            LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
                                       format(Ephemeris.datetimeToStr(currDt)))
                             rv.append(currDt)
 
                         elif prevDiff > 240 and currDiff < 120:
-                            LookbackMultipleUtils.log.debug("Passed the desired number of " + \
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Passed the desired number of " + \
                                       "elapsed degrees from " + \
                                       "above to below.  " + \
                                       "Narrowing down to the exact moment " + \
@@ -1013,7 +1058,8 @@ class LookbackMultipleUtils:
                             # Refine the timestamp until it is less
                             # than the threshold.
                             while currErrorTd > maxErrorTd:
-                                LookbackMultipleUtils.log.debug("Refining between {} and {}".\
+                                if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                    LookbackMultipleUtils.log.debug("Refining between {} and {}".\
                                           format(Ephemeris.datetimeToStr(t1),
                                                  Ephemeris.datetimeToStr(t2)))
                                 
@@ -1046,22 +1092,25 @@ class LookbackMultipleUtils:
                                 currErrorTd = Util.absTd(t2 - t1)
         
                             # currDt holds the moment in time.
-                            LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
+                            if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                                LookbackMultipleUtils.log.debug("Appending moment in time: {}".\
                                       format(Ephemeris.datetimeToStr(currDt)))
+
                             rv.append(currDt)
 
                         prevDt = currDt
-                        currDt = copy.deepcopy(currDt) + stepSizeTd
+                        currDt = currDt + stepSizeTd
                         prevDiff = currDiff
                         currDiff = None
 
-                    LookbackMultipleUtils.log.debug("Done searching for timestamps.")
+                    if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+                        LookbackMultipleUtils.log.debug("Done searching for timestamps.")
                         
                     # We have our timestamps, so we are done.
                     done = True
                     
             # Prepare for the next iteration.
-            steps.append(copy.deepcopy(steps[-1]) + stepSizeTd)
+            steps.append(steps[-1] + stepSizeTd)
             del steps[0]
             longitudesP1.append(None)
             del longitudesP1[0]
@@ -1069,7 +1118,8 @@ class LookbackMultipleUtils:
             # Update prevDiff as the currDiff.
             prevDiff = currDiff
 
-        LookbackMultipleUtils.log.debug("Exiting " + inspect.stack()[0][3] + "()")
+        if LookbackMultipleUtils.log.isEnabledFor(logging.DEBUG) == True:
+            LookbackMultipleUtils.log.debug("Exiting " + inspect.stack()[0][3] + "()")
         
         # The order of the datetimes in the list to be returned should be
         # chronological.  The way we extracted it was from later time to
@@ -1095,7 +1145,7 @@ class LookbackMultipleUtils:
                Ephemeris.isAscmcPlanetName(planetName):
 
             # House cusps and ascmc planets need a smaller step size.
-            stepSizeTd = datetime.timedelta(hours=4)
+            stepSizeTd = datetime.timedelta(hours=5)
 
         # These planets don't go retrograde, so we don't have to worry about
         # losing data points if our step size is too big.  We just have to keep
@@ -1104,330 +1154,32 @@ class LookbackMultipleUtils:
         # due to their elliptical orbits.
         if centricityType == "geocentric":
             if planetName == "Moon":
-                stepSizeTd = datetime.timedelta(days=3)
+                stepSizeTd = datetime.timedelta(days=5)
             elif planetName == "MoSu":
-                stepSizeTd = datetime.timedelta(days=3)
+                stepSizeTd = datetime.timedelta(days=5)
             elif planetName == "Sun":
-                stepSizeTd = datetime.timedelta(days=40)
+                stepSizeTd = datetime.timedelta(days=55)
         elif centricityType == "heliocentric":
             if planetName == "Mercury":
-                stepSizeTd = datetime.timedelta(days=3)
+                stepSizeTd = datetime.timedelta(days=12)
             elif planetName == "Venus":
-                stepSizeTd = datetime.timedelta(days=10)
+                stepSizeTd = datetime.timedelta(days=30)
             elif planetName == "Earth":
-                stepSizeTd = datetime.timedelta(days=40)
+                stepSizeTd = datetime.timedelta(days=55)
             elif planetName == "Mars":
-                stepSizeTd = datetime.timedelta(days=60)
+                stepSizeTd = datetime.timedelta(days=90)
             elif planetName == "Jupiter":
-                stepSizeTd = datetime.timedelta(days=120)
+                stepSizeTd = datetime.timedelta(days=360)
             elif planetName == "Saturn":
-                stepSizeTd = datetime.timedelta(days=180)
+                stepSizeTd = datetime.timedelta(days=720)
             elif planetName == "Uranus":
-                stepSizeTd = datetime.timedelta(days=240)
+                stepSizeTd = datetime.timedelta(days=2100)
             elif planetName == "Neptune":
-                stepSizeTd = datetime.timedelta(days=300)
+                stepSizeTd = datetime.timedelta(days=4200)
             elif planetName == "Pluto":
-                stepSizeTd = datetime.timedelta(days=400)
+                stepSizeTd = datetime.timedelta(days=4200)
 
         return stepSizeTd
-
-
-class LookbackMultiplePanelWidget(QWidget):
-    """Widget holding the QTableView that displays the PriceBar 
-    information along with other metrics analysis information.
-    """
-
-    # Signal emitted when the LookbackMultiples are modified by the user,
-    # either by checking a LookbackMultiple to be enabled, or by
-    # double-clicking and editing the particular LookbackMultiple in the list.
-    # Loading a new list of LookbackMultiples does not trigger this signal.
-    lookbackMultiplesModified = QtCore.pyqtSignal(list)
-
-    # Signal emitted when the "Apply/Redraw Lookback Multiples" button is
-    # clicked within this widget.
-    applyRedrawLookbackMultiplesButtonClicked = QtCore.pyqtSignal()
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        # Logger
-        self.log = \
-            logging.getLogger("lookbackmultiple.LookbackMultiplePanelWidget")
-
-        # Member variables.
-
-        # Holds the reference to the list of lookback multiples.
-        self.lookbackMultiples = []
-
-        self.applyRedrawButton = QPushButton("Apply/Redraw Lookback Multiples")
-
-        self.tableWidget = QTableWidget()
-        self.tableWidget.setColumnCount(1)
-        self.tableWidget.setRowCount(0)
-        self.tableWidget.setSortingEnabled(False)
-        self.tableWidget.setCornerButtonEnabled(False)
-        self.tableWidget.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.tableWidget.setSelectionMode(QAbstractItemView.NoSelection)
-        
-        self.tableWidgetHeaderItem = QTableWidgetItem("Lookback Multiples")
-        column = 0
-        self.tableWidget.setHorizontalHeaderItem(column, 
-                                                 self.tableWidgetHeaderItem)
-        self.tableWidget.horizontalHeader().setResizeMode(QHeaderView.Stretch)
-
-        # Setup the layout.
-        layout = QVBoxLayout()
-        layout.addWidget(self.applyRedrawButton)
-        layout.addWidget(self.tableWidget)
-        self.setLayout(layout)
-
-        # Connect signals and slots.
-        self.applyRedrawButton.clicked.\
-            connect(self.applyRedrawLookbackMultiplesButtonClicked)
-        self.tableWidget.itemChanged.\
-          connect(self._handleItemChanged)
-        self.tableWidget.itemDoubleClicked.\
-          connect(self._handleItemDoubleClicked)
-        
-        self.log.debug("size is: w={},h={}".\
-                       format(self.size().width(), 
-                              self.size().height()))
-        self.log.debug("sizeHint is: w={},h={}".\
-                       format(self.sizeHint().width(), 
-                              self.sizeHint().height()))
-
-    def getLookbackMultiples(self):
-        """Returns the reference to the list of LookbackMultiple objects."""
-        
-        if self.log.isEnabledFor(logging.DEBUG):
-            self.log.debug("Entered getLookbackMultiples()")
-
-            self.log.debug("len(self.lookbackMultiples) is currently: {}".\
-                           format(len(self.lookbackMultiples)))
-
-            self.log.debug("self.lookbackMultiples is currently: ".\
-                           format(self.lookbackMultiples))
-
-            for i in range(len(self.lookbackMultiples)):
-                lm = self.lookbackMultiples[i]
-                self.log.debug("self.lookbackMultiples[{}] is: {}".\
-                               format(i, lm.toString()))
-
-            self.log.debug("Exiting getLookbackMultiples()")
-
-        return self.lookbackMultiples
-
-
-    def setLookbackMultiples(self, lookbackMultiples):
-        """Sets the given list of LookbackMultiple objects to be
-        visualized in this widget.  This widget can potentially modify
-        the LookbackMultiple objects within this list.
-        """
-
-        self.log.debug("Entered setLookbackMultiples(lookbackMultiples)")
-
-        if lookbackMultiples == None:
-            self.log.error("Error: Invalid arguments.  " + \
-                           "lookbackMultiples to load cannot be None.")
-            self.log.debug("Exiting loadLookackMultiples(lookbackMultiples)")
-            return
-
-
-        # Here we will disconnect the signal-to-slot for handling when an item
-        # is changed.  We will reconnect it at the end of this method after we
-        # have made our modifications and created/modified the
-        # QTableWidgetItems within the QTableWidget.
-        # 
-        # Reason: This is because we don't want the signal to propagate
-        # upwards to indicate a real change has occured via user interaction
-        # from just loading.  
-        self.tableWidget.itemChanged.disconnect(self._handleItemChanged)
-
-        self.log.debug("Old lookbackMultiples has length: {}".\
-                       format(len(self.lookbackMultiples)))
-        self.log.debug("New lookbackMultiples has length: {}".\
-                       format(len(lookbackMultiples)))
-        self.log.debug("Before setting new lookbackMultiples, " + 
-                       "the number of rows is: {}".\
-                       format(self.tableWidget.rowCount()))
-     
-        # Store a reference to the new LookbackMultiples.
-        self.lookbackMultiples = lookbackMultiples
-
-        # Update the row count.
-        if self.tableWidget.rowCount() != len(self.lookbackMultiples):
-            self.tableWidget.setRowCount(len(self.lookbackMultiples))
-        
-        for i in range(len(self.lookbackMultiples)):
-            lookbackMultiple = lookbackMultiples[i]
-            row = i
-            column = 0
-            
-            # Get the existing item at the current row and column.  
-            # If it doesn't exist, then create it.
-            item = self.tableWidget.item(row, column)
-            if item == None:
-                self.log.debug("Item at (row={}, col={})".\
-                               format(row, column) + \
-                               " was None, so creating it now.")
-                item = QTableWidgetItem()
-                item.setFlags(Qt.ItemIsUserCheckable | Qt.ItemIsEnabled)
-                self.tableWidget.setItem(row, column, item)
-
-            # Set the internals of the QTableWidgetItem to reflect the
-            # LookbackMultiple that it represents.
-            if lookbackMultiple.getEnabled():
-                item.setCheckState(Qt.Checked)
-            else:
-                item.setCheckState(Qt.Unchecked)
-                
-            item.setIcon(ColorIcon(lookbackMultiple.getColor()))
-            item.setText(lookbackMultiple.toShortString())
-            item.setToolTip(\
-                self._getToolTipTextTextForLookbackMultiple(lookbackMultiple))
-            
-        # Reconnect the signal-to-slot, so that user changes to the items will
-        # notify us so we can update the internal LookbackMultiple variable(s), 
-        # and take any other necessary actions.
-        self.tableWidget.itemChanged.connect(self._handleItemChanged)
-
-        self.log.debug("Exiting setLookbackMultiples(lookbackMultiples)")
-
-    
-    def _handleItemChanged(self, item):
-        """Slot that handles when a QTableWidgetItem changes.
-        An example of a change that this handles is the item's check state.
-        
-        Arguments:
-        item - QTableWidgetItem that is sent via Qt's signal to this slot.
-        """
-        
-        self.log.debug("Entered _handleItemChanged(item)")
-
-        self.log.debug("Item changed at (row={}, col={})".\
-                       format(item.row(), item.column()))
-
-        # Currently, the only kinds of changes supported are checking the
-        # item's state, which will cause the enable state of the underlying
-        # LookbackMultiple to be updated.  The other way is to double-click the
-        # item to edit it manually, which another method handles this case.
-        row = item.row()
-        enabledFlag = item.checkState() == Qt.Checked
-        if self.lookbackMultiples[row].getEnabled() != enabledFlag:
-            self.log.debug("Checkstate was changed to {}.".format(enabledFlag))
-            self.lookbackMultiples[row].setEnabled(enabledFlag)
-
-            # Emit that there were modifications to the LoobkackMultiples.
-            self.lookbackMultiplesModified.emit(self.lookbackMultiples)
-        
-        self.log.debug("Exiting _handleItemChanged(item)")
-        
-    def _handleItemDoubleClicked(self, item):
-        """Slot that handles when a QTableWidgetItem is double-clicked.
-        This will open up an edit dialog to edit the underlying
-        LookbackMultiple.
-
-        Arguments:
-        item - QTableWidgetItem that is sent via Qt's signal to this slot.
-        """
-        
-        self.log.debug("Entered _handleItemDoubleClicked(item)")
-        
-        self.log.debug("Item double-clicked at (row={}, col={})".\
-                       format(item.row(), item.column()))
-                       
-        # Get the LookbackMultiple that underlies this QTableWidgetItem that
-        # was double-clicked.
-        row = item.row()
-        lookbackMultiple = self.lookbackMultiples[row]
-        
-        # Create a dialog to edit this LookbackMultiple.
-        dialog = LookbackMultipleEditDialog(lookbackMultiple)
-        
-        rv = dialog.exec_()
-        if rv == QDialog.Accepted:
-            # Dialog was accepted.  Obtain and store the new LookbackMultiple.
-            self.log.debug("Edit dialog for the LookbackMultiple was accepted.")
-            self.log.debug("Storing new LookbackMultiple back into the list.")
-
-            self.lookbackMultiples[row] = dialog.getLookbackMultiple()
-            lookbackMultiple = self.lookbackMultiples[row]
-
-            # Reload the current QTableWidgetItem so that it displays updated
-            # information.
-            # 
-            # (Here we will disconnect the itemChanged signal and then after
-            # updates, reconnect it).
-            self.tableWidget.itemChanged.disconnect(self._handleItemChanged)
-
-            if lookbackMultiple.getEnabled():
-                item.setCheckState(Qt.Checked)
-            else:
-                item.setCheckState(Qt.Unchecked)
-            item.setIcon(ColorIcon(lookbackMultiple.getColor()))
-            item.setText(lookbackMultiple.toShortString())
-            item.setToolTip(\
-                self._getToolTipTextTextForLookbackMultiple(lookbackMultiple))
-
-            self.tableWidget.itemChanged.connect(self._handleItemChanged)
-
-
-            # Emit that there were modifications to the LookbackMultiples.
-            self.lookbackMultiplesModified.emit(self.lookbackMultiples)
-        else:
-            self.log.debug("Edit dialog for the LookbackMultiple was rejected.")
-        
-        self.log.debug("Exiting _handleItemDoubleClicked(item)")
-        
-    def _getToolTipTextTextForLookbackMultiple(self, lookbackMultiple):
-        """Returns a formatted string to use for the tooltip."""
-        
-        lm = lookbackMultiple
-        
-        # Newlines.
-        endl = os.linesep
-
-        # Assemble substrings used within the final string.
-        centricityTypeStr = ""
-        if lm.getGeocentricFlag() == True:
-            centricityTypeStr = "G."
-        if lm.getHeliocentricFlag() == True:
-            centricityTypeStr = "H."
-            
-        planetNameStr = lm.getPlanetName()
-        lookbackMultipleStr = "{}".format(lm.getLookbackMultiple())
-        
-        longitudeTypeStr = ""
-        if lm.getTropicalFlag() == True:
-            longitudeTypeStr = "Trop."
-        if lm.getSiderealFlag() == True:
-            longitudeTypeStr = "Sid."
-            
-        baseUnitStr = "{}".format(lm.getBaseUnit())
-
-        baseUnitTypeStr = ""
-        if lm.getBaseUnitTypeDegreesFlag() == True:
-            baseUnitTypeStr = "degrees"
-        if lm.getBaseUnitTypeRevolutionsFlag() == True:
-            baseUnitTypeStr = "revolutions"
-
-        # Build the final string used for the tooltip.
-        formattedText = ""
-        formattedText += "Name: {}".format(lm.getName()) + endl
-        formattedText += "Enabled: {}".format(lm.getEnabled()) + endl
-        formattedText += \
-            "Summary: {}{} {} {} x {} {}".\
-            format(centricityTypeStr, 
-                   planetNameStr, 
-                   longitudeTypeStr,
-                   lookbackMultipleStr, 
-                   baseUnitStr, 
-                   baseUnitTypeStr) + endl
-        formattedText += "Description: {}".format(lm.getDescription())
-
-        return formattedText
-
-
 
 ##############################################################################
 
@@ -1465,7 +1217,6 @@ def testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInFuture():
     print("Running " + inspect.stack()[0][3] + "()")
     
     # Assumes that Ephemeris has been initialized by this point of execution.
-
 
     eastern = pytz.timezone('US/Eastern')
 
@@ -2132,97 +1883,63 @@ def testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInPast():
     print("")
 
 
-def testLookbackMultiplePanelWidget():
+def testLookbackMultipleUtils_speedTest():
+    """Tests to see how long it takes to do some computations."""
+
     print("Running " + inspect.stack()[0][3] + "()")
     
-    lm1 = LookbackMultiple(name="49ers",
-                          description="MyDescription1",
-                          lookbackMultiple=1.0,
-                          baseUnit=49.0,
-                          baseUnitTypeDegreesFlag=False,
-                          baseUnitTypeRevolutionsFlag=True,
-                          color=QColor(Qt.gray),
-                          enabled=False,
-                          planetName="Ascendant",
-                          geocentricFlag=True,
-                          heliocentricFlag=False,
-                          tropicalFlag=True,
-                          siderealFlag=False)
+    # For timing the calculations.
+    import time
 
-    lm2 = LookbackMultiple(name="40 deg Earth",
-                          description="MyDescription2",
-                          lookbackMultiple=1.0,
-                          baseUnit=40.0,
-                          baseUnitTypeDegreesFlag=True,
-                          baseUnitTypeRevolutionsFlag=False,
-                          color=QColor(Qt.red),
-                          enabled=True,
-                          planetName="Earth",
-                          geocentricFlag=False,
-                          heliocentricFlag=True,
-                          tropicalFlag=True,
-                          siderealFlag=False)
+    if True:
+        maxErrorTd = datetime.timedelta(minutes=60)
+        #maxErrorTd = datetime.timedelta(minutes=5)
+        #maxErrorTd = datetime.timedelta(seconds=2)
+        
+        print("  Testing G.AsSu moving 360 rev., 3 times, with maxErrorTd={}".\
+              format(maxErrorTd))
+        
+        startTime = time.time()
+        
+        for i in range(3):
+            planetName="AsSu"
+            centricityType="geocentric"
+            longitudeType="tropical"
+            referenceDt = datetime.datetime(1994, 10, 20, 0, 0, tzinfo=pytz.utc)
+            desiredDeltaDegrees = -360 * 360
 
-    lm3 = LookbackMultiple(name="360 deg Moon",
-                          description="MyDescription3",
-                          lookbackMultiple=1.0,
-                          baseUnit=360.0,
-                          baseUnitTypeDegreesFlag=True,
-                          baseUnitTypeRevolutionsFlag=False,
-                          color=QColor(Qt.blue),
-                          enabled=True,
-                          planetName="Moon",
-                          geocentricFlag=True,
-                          heliocentricFlag=False,
-                          tropicalFlag=True,
-                          siderealFlag=False)
+            resultDts = \
+                LookbackMultipleUtils.getDatetimesOfLongitudeDeltaDegreesInPast(
+                    planetName, centricityType, longitudeType, referenceDt, 
+                    desiredDeltaDegrees, maxErrorTd)
 
-    lm4 = LookbackMultiple(name="360 deg G.MoSu",
-                          description="MyDescription4",
-                          lookbackMultiple=1.0,
-                          baseUnit=360.0,
-                          baseUnitTypeDegreesFlag=True,
-                          baseUnitTypeRevolutionsFlag=False,
-                          color=QColor(Qt.black),
-                          enabled=True,
-                          planetName="MoSu",
-                          geocentricFlag=True,
-                          heliocentricFlag=False,
-                          tropicalFlag=True,
-                          siderealFlag=False)
+        endTime = time.time()
+        print("    Calculations took: {} sec".format(endTime - startTime))
 
-    lookbackMultiples = [lm1, lm2, lm3, lm4]
+    if True:
+        maxErrorTd = datetime.timedelta(minutes=60)
+        #maxErrorTd = datetime.timedelta(minutes=5)
+        #maxErrorTd = datetime.timedelta(seconds=2)
+        
+        print("  Testing G.MoSu moving 360 rev., 3 times, with maxErrorTd={}".\
+              format(maxErrorTd))
+        
+        startTime = time.time()
+        
+        for i in range(3):
+            planetName="MoSu"
+            centricityType="geocentric"
+            longitudeType="tropical"
+            referenceDt = datetime.datetime(1994, 10, 20, 0, 0, tzinfo=pytz.utc)
+            desiredDeltaDegrees = -360 * 360
 
-    widget = LookbackMultiplePanelWidget()
-    widget.setLookbackMultiples(lookbackMultiples)
+            resultDts = \
+                LookbackMultipleUtils.getDatetimesOfLongitudeDeltaDegreesInPast(
+                    planetName, centricityType, longitudeType, referenceDt, 
+                    desiredDeltaDegrees, maxErrorTd)
 
-    layout = QVBoxLayout()
-    layout.addWidget(widget)
-    
-    dialog = QDialog()
-    dialog.setLayout(layout)
-    rv = dialog.exec_()
-
-
-def testLookbackMultiplePanelWidgetEmpty():
-    print("Running " + inspect.stack()[0][3] + "()")
-    
-    # Just display the panel widget without any LookbackMultiples.
-    # This is to get an idea about the minimum/default sizes of the widget.
-    
-    lookbackMultiples = []
-
-    widget = LookbackMultiplePanelWidget()
-    #widget.setLookbackMultiples(lookbackMultiples)
-
-    layout = QVBoxLayout()
-    layout.addWidget(widget)
-    
-    dialog = QDialog()
-    dialog.setLayout(layout)
-    rv = dialog.exec_()
-
-    
+        endTime = time.time()
+        print("    Calculations took: {} sec".format(endTime - startTime))
 
 ##############################################################################
 
@@ -2230,6 +1947,12 @@ def testLookbackMultiplePanelWidgetEmpty():
 if __name__=="__main__":
     # For inspect.stack().
     import inspect
+
+    # For profiling.
+    #import cProfile
+
+    # For timing the calculations.
+    import time
 
     # For logging and for exiting.
     import os
@@ -2250,18 +1973,25 @@ if __name__=="__main__":
     logging.config.fileConfig(LOG_CONFIG_FILE)
 
     # Create the Qt application.
-    app = QApplication(sys.argv)
-
+    #app = QApplication(sys.argv)
 
     # Various tests to run:
-    testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInFuture()
-    testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInPast()
-    #testLookbackMultiplePanelWidget()
-    #testLookbackMultiplePanelWidgetEmpty()
 
+    def runTests():
+        testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInFuture()
+        testLookbackMultipleUtils_getDatetimesOfLongitudeDeltaDegreesInPast()
+        #testLookbackMultipleUtils_speedTest()
+
+    startTime = time.time()
+    runTests()
+    endTime = time.time()
+    print("Tests took: {} sec".format(endTime - startTime))
+
+    #cProfile.run('runTests()')
+    
     # Exit the app when all windows are closed.
-    app.connect(app, SIGNAL("lastWindowClosed()"), logging.shutdown)
-    app.connect(app, SIGNAL("lastWindowClosed()"), app, SLOT("quit()"))
+    #app.connect(app, SIGNAL("lastWindowClosed()"), logging.shutdown)
+    #app.connect(app, SIGNAL("lastWindowClosed()"), app, SLOT("quit()"))
 
     #app.exec_()
